CI: fix lint, coverage, and lint-win32 failures on master
Three independent CI breakages, all surfaced by the push of the
tooling + GoldSrc batches. Each was a real issue that didn't trip
local runs.

- src/pk3file.c: drop a redundant (size_t) cast on a value that's
  already size_t. clang-tidy's readability-redundant-casting check
  treats the cast as an error under -warnings-as-errors. Local
  Apple-toolchain clang-tidy doesn't ship this check by default;
  Ubuntu's does. Surfaced by the previous tooling-batch CI run too.

- Makefile coverage target: pass --ignore-errors unused to the
  lcov --remove call. lcov 2.x (Ubuntu's current package) treats
  "exclude pattern matched zero files" as a hard error. /usr/* and
  /Library/* are platform-specific noise that may or may not be
  present in the .info file depending on which host ran the build,
  so suppressing the unused-pattern check is the right call.
  Lcov 1.x (developer machines with older brew lcov) didn't enforce
  this so the bug only showed up in CI.

- Makefile lint-win32 target: switch -I src/vendor/{wingetopt,dirent}
  to -isystem so clang-tidy treats the vendored Mingw-compat shims
  as system headers. The default tidy check set flags reserved
  identifiers (__GETOPT_H__, _DIRENT_HAVE_D_TYPE, _wdirent, _WDIR,
  etc.) which the vendored sources use deliberately to match the
  POSIX shapes they mimic. -isystem suppresses diagnostics from
  those headers the same way the normal make lint already skips
  src/vendor/ via the $(SOURCES) glob. Documented inline.

Local validation:
- make test: 148/148 pass
- make CLANG_TIDY=$(brew --prefix llvm)/bin/clang-tidy lint: exits 0
- make lint-win32 / make REQUIRE_WIN32_HEADERS=1 lint-win32: same
  soft-skip / hard-fail behavior as before (real Mingw run still
  deferred to CI since local install is absent)

+++ src/pk3file.c
@@ -1812,8 +1812,7 @@ pakka_status_t pakka_pk3_commit_impl(Pak_t *pak, pakka_error_t *err) {
                                                 e->filename);
                         }
                         recomputed_crc = pk3_crc32_update(recomputed_crc,
-                                                         copy_buf,
-                                                         (size_t)got);
+                                                         copy_buf, got);
                         remaining -= (uint64_t)got;
                     }
                     fclose(src);
